backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
requests
gunicorn
drf-spectacular
pytest
pytest-django
coverage
